Fix : Review 모델 grade 필드 타입 변경 #25

+++ perfume/models.py
@@ -30,7 +30,7 @@ class Review(models.Model):
     perfume = models.ForeignKey(Perfume, on_delete=models.CASCADE, related_name="perfume_reviews")
     good_content = models.TextField()
     bad_content = models.TextField()
-    grade = models.IntegerField(default=5, null=False, blank=True)
+    grade = models.FloatField(default=5, null=False, blank=True)
     image = models.ImageField(null=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
